data-engineering/src/scheduler/eia_petroleum/requirements.txt
pandas==2.2.3
pydantic
PyYAML
requests
db-dtypes
google-cloud-bigquery
pymsteams
pandas_gbq
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
